+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added more useful information when JSON file is invalid (ie. your main config).
 * `pos-cli sync` will not stop if `template-values.json` is invalid JSON.
 * Improved error handling for templates exception.
+* Improved error messages returned by the server.
 
 ## 4.0.3 25 July, 2019
 * Init is now not overriding files in current directory.